only destroy information in development

+++ db/seeds.rb
@@ -2,7 +2,7 @@ require "open-uri"
 
 puts "Destroying the data"
 Doll.destroy_all if Rails.env.development?
-User.destroy_all
+User.destroy_all if Rails.env.development?
 puts "Creating new ones"
 
 user_1 = User.create!(
